Add postgres DB for joomla pr testing

+++ joomla-pr/scripts/exec-joomla-pr-webserver.sh
@@ -15,7 +15,7 @@ fi
 root=/var/www/html/pr/$1
 
 # Empty the install directory if rebuild is set
-if [ ! -z $2 ]; then
+if [ ! -z $3 ]; then
   echo "Cleaning install folder"
   sudo rm -rf $root
 fi
@@ -37,7 +37,7 @@ git fetch origin pull/$1/head:pr-$1 &>/dev/null
 git checkout pr-$1 &>/dev/null
 
 # Run composer
-if [ ! -z $2 ]; then
+if [ ! -z $3 ]; then
   rm -rf $root/libraries/vendor
 fi
 echo "Installing PHP dependencies"
@@ -45,7 +45,7 @@ composer install --quiet
 
 # Run npm
 if [ -f $root/package.json ]; then
-  if [ ! -z $2 ]; then
+  if [ ! -z $3 ]; then
     echo "Cleaning the assets"
     npm c &>/dev/null
   fi
@@ -57,19 +57,45 @@ fi
 
 # Install Joomla when no configuration file is available
 if [ ! -f $root/configuration.php ]; then
-  echo "Installing Joomla"
+  echo "Setup Joomla"
+
+  # Setup configuration file
   cp $(dirname $0)/configuration.php $root/configuration.php
   sed -i "s/{PR}/$1/g" $root/configuration.php
+  sed -i "s/{DBHOST}/$2-pr/g" $root/configuration.php
 
-  mysql -u root -proot -h mysql-pr -e "drop database if exists joomla_pr_$1";
-  mysql -u root -proot -h mysql-pr -e "create database joomla_pr_$1";
-  sed "s/#_/j/g" $root/installation/sql/mysql/joomla.sql | mysql -u root -proot -h mysql-pr -D joomla_pr_$1
-  mysql -u root -proot -h mysql-pr -D joomla_pr_$1 -e "INSERT INTO j_users (id, name, username, email, password, block) VALUES
-(42, 'admin', 'admin', 'admin@example.com', '\$2y\$10\$O.A8bZcuC6yFfgjzycqzku7LuG6zvBiozJcjXD4FP3bhJdvyKdtoG', 0);";
-  mysql -u root -proot -h mysql-pr -D joomla_pr_$1 -e "INSERT INTO j_user_usergroup_map (user_id, group_id) VALUES ('42', '8');";
+  if [[ -z $2 || $2 == 'mysql' ]]; then
+    echo "Installing Joomla on mysql"
+    sed -i "s/{DBDRIVER}/mysqli/g" $root/configuration.php
+
+    # Install joomla
+    mysql -u root -proot -h mysql-pr -e "drop database if exists joomla_pr_$1"
+    mysql -u root -proot -h mysql-pr -e "create database joomla_pr_$1"
+    sed "s/#_/j/g" $root/installation/sql/mysql/joomla.sql | mysql -u root -proot -h mysql-pr -D joomla_pr_$1
+    mysql -u root -proot -h mysql-pr -D joomla_pr_$1 -e "INSERT INTO j_users (id, name, username, email, password, block) VALUES(42, 'admin', 'admin', 'admin@example.com', '\$2y\$10\$O.A8bZcuC6yFfgjzycqzku7LuG6zvBiozJcjXD4FP3bhJdvyKdtoG', 0)"
+    mysql -u root -proot -h mysql-pr -D joomla_pr_$1 -e "INSERT INTO j_user_usergroup_map (user_id, group_id) VALUES ('42', '8')"
+  fi
+
+  if [[ $2 == 'postgres' ]]; then
+    echo "Installing Joomla on postgres"
+    sed -i "s/{DBDRIVER}/pgsql/g" $root/configuration.php
+    export PGPASSWORD=root
+
+    # Clear existing connections
+    psql -U root -h postgres-pr -c "REVOKE CONNECT ON DATABASE joomla_pr_$1 FROM public" > /dev/null 2>&1
+    psql -U root -h postgres-pr -c "SELECT pid, pg_terminate_backend(pid) FROM pg_stat_activity WHERE datname = 'joomla_pr_$1' AND pid <> pg_backend_pid()" > /dev/null
+
+    # Install joomla
+    psql -U root -h postgres-pr -c "drop database if exists joomla_pr_$1" > /dev/null
+    psql -U root -h postgres-pr -c "create database joomla_pr_$1" > /dev/null
+    sed "s/#_/j/g" $root/installation/sql/postgresql/joomla.sql | psql -U root -h postgres-pr -d joomla_pr_$1 > /dev/null
+    psql -U root -h postgres-pr -d joomla_pr_$1 -c "INSERT INTO j_users (id, name, username, email, password, block,  \"registerDate\", params) VALUES(42, 'admin', 'admin', 'admin@example.com', '\$2y\$10\$O.A8bZcuC6yFfgjzycqzku7LuG6zvBiozJcjXD4FP3bhJdvyKdtoG', 0, '2020-01-01 00:00:00', '{}')" > /dev/null
+    psql -U root -h postgres-pr -d joomla_pr_$1 -c "INSERT INTO j_user_usergroup_map (user_id, group_id) VALUES ('42', '8')" > /dev/null
+  fi
 fi
 
 echo -e "\e[32mPR $1 is checked out and ready to test on http://localhost:8090/pr/$1!"
 echo -e "\e[32mYou can log in with admin/admin!"
 echo -e "\e[32mPHPMyAdmin is available on http://localhost:8091!"
-echo -e "\e[32mMailcatcher is available on http://localhost:8092!"
+echo -e "\e[32mpgAdmin is available on http://localhost:8092, credentials are admin@example.com/root!"
+echo -e "\e[32mMailcatcher is available on http://localhost:8093!"
